A* scored by ints instead floats
Mostly trivial optimization improving 1-3%. Tried some attempts to reduce instruction counts of push/pop, but the savings were negligible for the extra complexity. The real cost is the number items being pushed to the priority queue which would require a larger change to improve.
Moving on to refactoring the RoomGeometry to separate the simulate methods and the "caches" which caused the mutable hacks.